Fetching data in Dashboard and not in ReceiptList
